pyproject.toml
[project]
name = "compliance-checker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-ai-documentintelligence>=1.0.2",
    "azure-identity>=1.25.2",
    "python-dotenv>=1.2.1",
]
